# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智慧物业管理系统 SaaS 多租户版 API
  version: 1.1.1
  description: |
    本接口契约用于约束 Codex 开发。实际开发不得擅自更改路径、核心字段、状态枚举和权限标识。
    所有业务接口默认返回 ApiResponse；列表接口 data 使用 PageResult。
servers:
  - url: https://api.example.com
    description: 生产环境示例
  - url: http://localhost:8080
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
...
  /api/base/members:
    get:
      tags: [Base]
      summary: 业主/住户列表
      x-permission: base:member:list
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: keyword
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
  /api/base/member-bindings:
    get:
      tags: [Base]
      summary: 房屋绑定审核列表
      x-permission: base:memberBinding:list
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: memberId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/BindStatus' }
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
  /api/base/member-bindings/{bindId}/audit:
    put:
      tags: [Base]
      summary: 审核房屋绑定
      x-permission: base:memberBinding:audit
      parameters:
        - $ref: '#/components/parameters/BindId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuditRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/base/member-bindings/{bindId}/unbind:
    put:
      tags: [Base]
      summary: 后台解绑房屋绑定
      x-permission: base:memberBinding:unbind
      parameters:
        - $ref: '#/components/parameters/BindId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
  /api/report/fee/summary:
    get:
      tags: [Report]
      summary: 收费汇总报表
      x-permission: report:fee:view
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/report/workorders/summary:
    get:
      tags: [Report]
      summary: 工单汇总报表
      x-permission: report:workorder:view
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/report/patrol/summary:
    get:
      tags: [Report]
      summary: 巡检汇总报表
      x-permission: report:patrol:view
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/report/lease/summary:
    get:
      tags: [Report]
      summary: 租赁汇总报表
      x-permission: report:lease:view
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
  /api/reports:
    get:
      tags: [Report]
      summary: 报表中心入口
      x-permission: report:center:view
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/app/auth/wx-login:
    post:
      tags: [App]
      summary: 微信小程序登录
      security: []
      x-permission: public
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WxLoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
  /api/app/house-bindings:
    post:
      tags: [App]
      summary: 提交房屋绑定申请
      x-permission: app:house:bind
      security: []
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HouseBindingApplyRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/app/house-bindings/{bindId}/unbind:
    put:
      tags: [App]
      summary: 小程序房屋解绑
      x-permission: app:house:unbind
      parameters:
        - name: bindId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnbindRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
...
    BindId:
      name: bindId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    ApiResponse:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    PageResponse:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/PageResult' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true }
        traceId: { type: string }
...
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: { type: object }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
...
    BindStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, UNBOUND, EXPIRED]
...
    WxLoginRequest:
      type: object
      required: [tenantId, openid]
      properties:
        tenantId: { type: integer, format: int64 }
        openid: { type: string }
        unionid: { type: string }
        mobile: { type: string }
        realName: { type: string }
        avatarUrl: { type: string }
    HouseBindingApplyRequest:
      type: object
      required: [tenantId, projectId, memberId, houseId, bindRole, realName, mobile]
      properties:
        tenantId: { type: integer, format: int64 }
        projectId: { type: integer, format: int64 }
        memberId: { type: integer, format: int64 }
        houseId: { type: integer, format: int64 }
        bindRole: { type: string, enum: [OWNER, FAMILY, TENANT, RESIDENT] }
        realName: { type: string }
        mobile: { type: string }
        idCardNoEncrypted: { type: string }
        proofFileIds: { type: string }
        effectiveDate: { type: string, format: date }
        expireDate: { type: string, format: date }
    UnbindRequest:
      type: object
      properties:
        reason: { type: string }
...
    AuditRequest:
      type: object
      required: [auditResult]
      properties:
        auditResult: { type: string, enum: [APPROVED, REJECTED] }
        auditRemark: { type: string }
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
...